feat: reorganize diagnostics in Settings (#11)
Move "Copy Logs" and "Open Console" from About tab to a new Diagnostics
section on the General tab with descriptive help text. Move "Check for
Updates…" button from General tab to About tab, alongside version info.
